Removed superfluous UTF8 test.
The test was lame, it checked the version number of the mysql database
to start with a 4, 5, 6, or 7.

In the mean time we have a minimum version of 5 that is needed for OTRS.

Also, this test breaks on MariaDB 10, because 10 was not in the list.

Apart from that we have _UTF8ClientCheck and _UTF8DatabaseCheck and it
if these tests succeed, then the database can do UTF8.